Crate wincolor

This crate provides a safe and simple Windows specific API to control text attributes in the Windows console. Text attributes are limited to foreground/background colors, as well as whether to make colors intense or not.

Note that on non-Windows platforms, this crate is empty but will compile.

Example

use wincolor::{Console, Color, Intense};

let mut con = Console::stdout().unwrap();
con.fg(Intense::Yes, Color::Cyan).unwrap();
println!("This text will be intense cyan.");
con.reset().unwrap();
println!("This text will be normal.");

Structs

Console

A Windows console.

Enums

Color

The set of available colors for use with a Windows console.

Intense

Whether to use intense colors or not.